rename folder, remove build package
