--- src/main/java/com/epam/jooq/DefaultCatalog.java
@@ -1,51 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.epam.jooq;
-
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import org.jooq.Schema;
-import org.jooq.impl.CatalogImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class DefaultCatalog extends CatalogImpl {
-
-    private static final long serialVersionUID = 1676713638;
-
-    /**
-     * The reference instance of <code></code>
-     */
-    public static final DefaultCatalog DEFAULT_CATALOG = new DefaultCatalog();
-
-    /**
-     * The schema <code>public</code>.
-     */
-    public final Public PUBLIC = com.epam.jooq.Public.PUBLIC;
-
-    /**
-     * No further instances allowed
-     */
-    private DefaultCatalog() {
-        super("");
-    }
-
-    @Override
-    public final List<Schema> getSchemas() {
-        List result = new ArrayList();
-        result.addAll(getSchemas0());
-        return result;
-    }
-
-    private final List<Schema> getSchemas0() {
-        return Arrays.<Schema>asList(
-            Public.PUBLIC);
-    }
-}
