--- src/main/java/com/epam/jooq/Indexes.java
@@ -1,40 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.epam.jooq;
-
-
-import com.epam.jooq.tables.Description;
-import com.epam.jooq.tables.Product;
-
-import org.jooq.Index;
-import org.jooq.OrderField;
-import org.jooq.impl.Internal;
-
-
-/**
- * A class modelling indexes of tables of the <code>public</code> schema.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Indexes {
-
-    // -------------------------------------------------------------------------
-    // INDEX definitions
-    // -------------------------------------------------------------------------
-
-    public static final Index DESCRIPTION_PKEY = Indexes0.DESCRIPTION_PKEY;
-    public static final Index PRODUCT_PKEY = Indexes0.PRODUCT_PKEY;
-    public static final Index VENDOR_INDEX = Indexes0.VENDOR_INDEX;
-    public static final Index VENDOR_NAME = Indexes0.VENDOR_NAME;
-
-    // -------------------------------------------------------------------------
-    // [#1459] distribute members to avoid static initialisers > 64kb
-    // -------------------------------------------------------------------------
-
-    private static class Indexes0 {
-        public static Index DESCRIPTION_PKEY = Internal.createIndex("description_pkey", Description.DESCRIPTION, new OrderField[] { Description.DESCRIPTION.ID }, true);
-        public static Index PRODUCT_PKEY = Internal.createIndex("product_pkey", Product.PRODUCT, new OrderField[] { Product.PRODUCT.ID }, true);
-        public static Index VENDOR_INDEX = Internal.createIndex("vendor_index", Product.PRODUCT, new OrderField[] { Product.PRODUCT.NAME }, true);
-        public static Index VENDOR_NAME = Internal.createIndex("vendor_name", Product.PRODUCT, new OrderField[] { Product.PRODUCT.NAME }, true);
-    }
-}
