--- src/main/java/com/epam/jooq/Keys.java
@@ -1,56 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.epam.jooq;
-
-
-import com.epam.jooq.tables.Description;
-import com.epam.jooq.tables.Product;
-import com.epam.jooq.tables.records.DescriptionRecord;
-import com.epam.jooq.tables.records.ProductRecord;
-
-import org.jooq.Identity;
-import org.jooq.UniqueKey;
-import org.jooq.impl.Internal;
-
-
-/**
- * A class modelling foreign key relationships and constraints of tables of 
- * the <code>public</code> schema.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Keys {
-
-    // -------------------------------------------------------------------------
-    // IDENTITY definitions
-    // -------------------------------------------------------------------------
-
-    public static final Identity<DescriptionRecord, Long> IDENTITY_DESCRIPTION = Identities0.IDENTITY_DESCRIPTION;
-    public static final Identity<ProductRecord, Integer> IDENTITY_PRODUCT = Identities0.IDENTITY_PRODUCT;
-
-    // -------------------------------------------------------------------------
-    // UNIQUE and PRIMARY KEY definitions
-    // -------------------------------------------------------------------------
-
-    public static final UniqueKey<DescriptionRecord> DESCRIPTION_PKEY = UniqueKeys0.DESCRIPTION_PKEY;
-    public static final UniqueKey<ProductRecord> PRODUCT_PKEY = UniqueKeys0.PRODUCT_PKEY;
-
-    // -------------------------------------------------------------------------
-    // FOREIGN KEY definitions
-    // -------------------------------------------------------------------------
-
-
-    // -------------------------------------------------------------------------
-    // [#1459] distribute members to avoid static initialisers > 64kb
-    // -------------------------------------------------------------------------
-
-    private static class Identities0 {
-        public static Identity<DescriptionRecord, Long> IDENTITY_DESCRIPTION = Internal.createIdentity(Description.DESCRIPTION, Description.DESCRIPTION.ID);
-        public static Identity<ProductRecord, Integer> IDENTITY_PRODUCT = Internal.createIdentity(Product.PRODUCT, Product.PRODUCT.ID);
-    }
-
-    private static class UniqueKeys0 {
-        public static final UniqueKey<DescriptionRecord> DESCRIPTION_PKEY = Internal.createUniqueKey(Description.DESCRIPTION, "description_pkey", Description.DESCRIPTION.ID);
-        public static final UniqueKey<ProductRecord> PRODUCT_PKEY = Internal.createUniqueKey(Product.PRODUCT, "product_pkey", Product.PRODUCT.ID);
-    }
-}
