--- src/main/java/com/epam/jooq/Public.java
@@ -1,67 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.epam.jooq;
-
-
-import com.epam.jooq.tables.Description;
-import com.epam.jooq.tables.Product;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import org.jooq.Catalog;
-import org.jooq.Table;
-import org.jooq.impl.SchemaImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Public extends SchemaImpl {
-
-    private static final long serialVersionUID = 226925495;
-
-    /**
-     * The reference instance of <code>public</code>
-     */
-    public static final Public PUBLIC = new Public();
-
-    /**
-     * The table <code>public.description</code>.
-     */
-    public final Description DESCRIPTION = com.epam.jooq.tables.Description.DESCRIPTION;
-
-    /**
-     * The table <code>public.product</code>.
-     */
-    public final Product PRODUCT = com.epam.jooq.tables.Product.PRODUCT;
-
-    /**
-     * No further instances allowed
-     */
-    private Public() {
-        super("public", null);
-    }
-
-
-    @Override
-    public Catalog getCatalog() {
-        return DefaultCatalog.DEFAULT_CATALOG;
-    }
-
-    @Override
-    public final List<Table<?>> getTables() {
-        List result = new ArrayList();
-        result.addAll(getTables0());
-        return result;
-    }
-
-    private final List<Table<?>> getTables0() {
-        return Arrays.<Table<?>>asList(
-            Description.DESCRIPTION,
-            Product.PRODUCT);
-    }
-}
