--- src/main/java/com/epam/jooq/Tables.java
@@ -1,26 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.epam.jooq;
-
-
-import com.epam.jooq.tables.Description;
-import com.epam.jooq.tables.Product;
-
-
-/**
- * Convenience access to all tables in public
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Tables {
-
-    /**
-     * The table <code>public.description</code>.
-     */
-    public static final Description DESCRIPTION = Description.DESCRIPTION;
-
-    /**
-     * The table <code>public.product</code>.
-     */
-    public static final Product PRODUCT = Product.PRODUCT;
-}
