--- src/main/java/com/epam/jooq/tables/Description.java
@@ -1,158 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.epam.jooq.tables;
-
-
-import com.epam.jooq.Indexes;
-import com.epam.jooq.Keys;
-import com.epam.jooq.Public;
-import com.epam.jooq.tables.records.DescriptionRecord;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.jooq.Field;
-import org.jooq.ForeignKey;
-import org.jooq.Identity;
-import org.jooq.Index;
-import org.jooq.Name;
-import org.jooq.Record;
-import org.jooq.Row3;
-import org.jooq.Schema;
-import org.jooq.Table;
-import org.jooq.TableField;
-import org.jooq.UniqueKey;
-import org.jooq.impl.DSL;
-import org.jooq.impl.TableImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Description extends TableImpl<DescriptionRecord> {
-
-    private static final long serialVersionUID = 1273515235;
-
-    /**
-     * The reference instance of <code>public.description</code>
-     */
-    public static final Description DESCRIPTION = new Description();
-
-    /**
-     * The class holding records for this type
-     */
-    @Override
-    public Class<DescriptionRecord> getRecordType() {
-        return DescriptionRecord.class;
-    }
-
-    /**
-     * The column <code>public.description.id</code>.
-     */
-    public final TableField<DescriptionRecord, Long> ID = createField(DSL.name("id"), org.jooq.impl.SQLDataType.BIGINT.nullable(false).defaultValue(org.jooq.impl.DSL.field("nextval('description_id_seq'::regclass)", org.jooq.impl.SQLDataType.BIGINT)), this, "");
-
-    /**
-     * The column <code>public.description.product_id</code>.
-     */
-    public final TableField<DescriptionRecord, Integer> PRODUCT_ID = createField(DSL.name("product_id"), org.jooq.impl.SQLDataType.INTEGER.nullable(false), this, "");
-
-    /**
-     * The column <code>public.description.descr</code>.
-     */
-    public final TableField<DescriptionRecord, Integer> DESCR = createField(DSL.name("descr"), org.jooq.impl.SQLDataType.INTEGER.nullable(false), this, "");
-
-    /**
-     * Create a <code>public.description</code> table reference
-     */
-    public Description() {
-        this(DSL.name("description"), null);
-    }
-
-    /**
-     * Create an aliased <code>public.description</code> table reference
-     */
-    public Description(String alias) {
-        this(DSL.name(alias), DESCRIPTION);
-    }
-
-    /**
-     * Create an aliased <code>public.description</code> table reference
-     */
-    public Description(Name alias) {
-        this(alias, DESCRIPTION);
-    }
-
-    private Description(Name alias, Table<DescriptionRecord> aliased) {
-        this(alias, aliased, null);
-    }
-
-    private Description(Name alias, Table<DescriptionRecord> aliased, Field<?>[] parameters) {
-        super(alias, null, aliased, parameters, DSL.comment(""));
-    }
-
-    public <O extends Record> Description(Table<O> child, ForeignKey<O, DescriptionRecord> key) {
-        super(child, key, DESCRIPTION);
-    }
-
-    @Override
-    public Schema getSchema() {
-        return Public.PUBLIC;
-    }
-
-    @Override
-    public List<Index> getIndexes() {
-        return Arrays.<Index>asList(Indexes.DESCRIPTION_PKEY);
-    }
-
-    @Override
-    public Identity<DescriptionRecord, Long> getIdentity() {
-        return Keys.IDENTITY_DESCRIPTION;
-    }
-
-    @Override
-    public UniqueKey<DescriptionRecord> getPrimaryKey() {
-        return Keys.DESCRIPTION_PKEY;
-    }
-
-    @Override
-    public List<UniqueKey<DescriptionRecord>> getKeys() {
-        return Arrays.<UniqueKey<DescriptionRecord>>asList(Keys.DESCRIPTION_PKEY);
-    }
-
-    @Override
-    public Description as(String alias) {
-        return new Description(DSL.name(alias), this);
-    }
-
-    @Override
-    public Description as(Name alias) {
-        return new Description(alias, this);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Description rename(String name) {
-        return new Description(DSL.name(name), null);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Description rename(Name name) {
-        return new Description(name, null);
-    }
-
-    // -------------------------------------------------------------------------
-    // Row3 type methods
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row3<Long, Integer, Integer> fieldsRow() {
-        return (Row3) super.fieldsRow();
-    }
-}
