--- src/main/java/com/epam/jooq/tables/Product.java
@@ -1,179 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.epam.jooq.tables;
-
-
-import com.epam.jooq.Indexes;
-import com.epam.jooq.Keys;
-import com.epam.jooq.Public;
-import com.epam.jooq.tables.records.ProductRecord;
-
-import java.math.BigDecimal;
-import java.util.Arrays;
-import java.util.List;
-
-import org.jooq.Field;
-import org.jooq.ForeignKey;
-import org.jooq.Identity;
-import org.jooq.Index;
-import org.jooq.Name;
-import org.jooq.Record;
-import org.jooq.Row7;
-import org.jooq.Schema;
-import org.jooq.Table;
-import org.jooq.TableField;
-import org.jooq.UniqueKey;
-import org.jooq.impl.DSL;
-import org.jooq.impl.TableImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Product extends TableImpl<ProductRecord> {
-
-    private static final long serialVersionUID = 98946542;
-
-    /**
-     * The reference instance of <code>public.product</code>
-     */
-    public static final Product PRODUCT = new Product();
-
-    /**
-     * The class holding records for this type
-     */
-    @Override
-    public Class<ProductRecord> getRecordType() {
-        return ProductRecord.class;
-    }
-
-    /**
-     * The column <code>public.product.id</code>.
-     */
-    public final TableField<ProductRecord, Integer> ID = createField(DSL.name("id"), org.jooq.impl.SQLDataType.INTEGER.nullable(false).defaultValue(org.jooq.impl.DSL.field("nextval('product_id_seq'::regclass)", org.jooq.impl.SQLDataType.INTEGER)), this, "");
-
-    /**
-     * The column <code>public.product.name</code>.
-     */
-    public final TableField<ProductRecord, String> NAME = createField(DSL.name("name"), org.jooq.impl.SQLDataType.CLOB.nullable(false), this, "");
-
-    /**
-     * The column <code>public.product.price</code>.
-     */
-    public final TableField<ProductRecord, BigDecimal> PRICE = createField(DSL.name("price"), org.jooq.impl.SQLDataType.NUMERIC, this, "");
-
-    /**
-     * The column <code>public.product.category_id</code>.
-     */
-    public final TableField<ProductRecord, Integer> CATEGORY_ID = createField(DSL.name("category_id"), org.jooq.impl.SQLDataType.INTEGER.nullable(false), this, "");
-
-    /**
-     * The column <code>public.product.vendor</code>.
-     */
-    public final TableField<ProductRecord, String> VENDOR = createField(DSL.name("vendor"), org.jooq.impl.SQLDataType.CLOB, this, "");
-
-    /**
-     * The column <code>public.product.sn</code>.
-     */
-    public final TableField<ProductRecord, String> SN = createField(DSL.name("sn"), org.jooq.impl.SQLDataType.CLOB, this, "");
-
-    /**
-     * The column <code>public.product.owner</code>.
-     */
-    public final TableField<ProductRecord, String> OWNER = createField(DSL.name("owner"), org.jooq.impl.SQLDataType.CLOB, this, "");
-
-    /**
-     * Create a <code>public.product</code> table reference
-     */
-    public Product() {
-        this(DSL.name("product"), null);
-    }
-
-    /**
-     * Create an aliased <code>public.product</code> table reference
-     */
-    public Product(String alias) {
-        this(DSL.name(alias), PRODUCT);
-    }
-
-    /**
-     * Create an aliased <code>public.product</code> table reference
-     */
-    public Product(Name alias) {
-        this(alias, PRODUCT);
-    }
-
-    private Product(Name alias, Table<ProductRecord> aliased) {
-        this(alias, aliased, null);
-    }
-
-    private Product(Name alias, Table<ProductRecord> aliased, Field<?>[] parameters) {
-        super(alias, null, aliased, parameters, DSL.comment(""));
-    }
-
-    public <O extends Record> Product(Table<O> child, ForeignKey<O, ProductRecord> key) {
-        super(child, key, PRODUCT);
-    }
-
-    @Override
-    public Schema getSchema() {
-        return Public.PUBLIC;
-    }
-
-    @Override
-    public List<Index> getIndexes() {
-        return Arrays.<Index>asList(Indexes.PRODUCT_PKEY, Indexes.VENDOR_INDEX, Indexes.VENDOR_NAME);
-    }
-
-    @Override
-    public Identity<ProductRecord, Integer> getIdentity() {
-        return Keys.IDENTITY_PRODUCT;
-    }
-
-    @Override
-    public UniqueKey<ProductRecord> getPrimaryKey() {
-        return Keys.PRODUCT_PKEY;
-    }
-
-    @Override
-    public List<UniqueKey<ProductRecord>> getKeys() {
-        return Arrays.<UniqueKey<ProductRecord>>asList(Keys.PRODUCT_PKEY);
-    }
-
-    @Override
-    public Product as(String alias) {
-        return new Product(DSL.name(alias), this);
-    }
-
-    @Override
-    public Product as(Name alias) {
-        return new Product(alias, this);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Product rename(String name) {
-        return new Product(DSL.name(name), null);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Product rename(Name name) {
-        return new Product(name, null);
-    }
-
-    // -------------------------------------------------------------------------
-    // Row7 type methods
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row7<Integer, String, BigDecimal, Integer, String, String, String> fieldsRow() {
-        return (Row7) super.fieldsRow();
-    }
-}
